==> ranking/0 - gathering data/2 - concat dataframes.ipynb <==
import pandas as pd

dataframes = []
for i in range(5):
    dataframes.append(pd.read_csv(f"glassfish_{i}.csv"))
df = pd.concat(dataframes, ignore_index=True)
print(df.shape)
df.to_csv("glassfish_0.csv", index=False)
